python-scripts/analysis/requirements.txt
tensorflow-gpu
pandas
sklearn
numpy